+++ Kernel/System/Ticket/Event/NotificationEvent/Transport/Email.pm
@@ -9,6 +9,7 @@
 package Kernel::System::Ticket::Event::NotificationEvent::Transport::Email;
 ## nofilter(TidyAll::Plugin::OTRS::Perl::LayoutObject)
 ## nofilter(TidyAll::Plugin::OTRS::Perl::ParamObject)
+## nofilter(TidyAll::Plugin::OTRS::Perl::ObjectManagerCreation)
 
 use strict;
 use warnings;
@@ -138,6 +139,16 @@ sub SendNotification {
 
         if ( !-r "$TemplateDir/$EmailTemplate.tt" && !-r "$CustomTemplateDir/$EmailTemplate.tt" ) {
             $EmailTemplate = 'Default';
+        }
+
+        # Create local OM with modified Layout object to include recipient language preference if applicable.
+        if ( $Recipient{UserLanguage} ) {
+            local $Kernel::OM = Kernel::System::ObjectManager->new(
+                'Kernel::Output::HTML::Layout' => {
+                    Lang => $Recipient{UserLanguage},
+                },
+            );
+            $LayoutObject = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
         }
 
         # generate HTML